Allow user-added listeners to block the exit
If a user adds a SIGHUP listener, then they can respond to that SIGHUP
and either process.exit() or not.  With this patch, that behavior is
restored.

+++ index.js
@@ -7,10 +7,12 @@ module.exports = function (cb) {
 
   Object.keys(signals).forEach(function (sig) {
     var listener = function () {
-      process.removeListener(sig, listener)
-      cb(process.exitCode || signals[sig], sig)
-
-      process.kill(process.pid, sig)
+      // If there are no other listeners, do the default action.
+      if (process.listeners(sig).length === 1) {
+        process.removeListener(sig, listener)
+        cb(process.exitCode || signals[sig], sig)
+        process.kill(process.pid, sig)
+      }
     }
 
     try {
